Refactor components, update imports, add new component usages
