--- apps/web/constants/Sidebar.tsx
@@ -1,72 +1,0 @@
-import {
-  Title as TableTitle,
-  href as TableHref,
-} from "@repo/components/core/Table";
-import {
-  Title as PaginationTitle,
-  href as PaginationHref,
-} from "@repo/components/core/Pagination";
-import {
-  Title as ToastTitle,
-  href as ToastHref,
-} from "@repo/components/core/Toast";
-import {
-  Title as CheckboxTitle,
-  href as CheckboxHref,
-} from "@repo/components/core/Checkbox";
-
-import {
-  Title as DialogTitle,
-  href as DialogHref,
-} from "@repo/components/core/DialogBox";
-
-import {
-  Title as SwitchTitle,
-  href as SwitchHref,
-} from "@repo/components/core/Switch";
-
-import {
-  Title as TooltipTitle,
-  href as TooltipHref,
-} from "@repo/components/core/Tooltip";
-
-export const Reference = [
-  {
-    groupKey: "Core",
-    groupTitle: "Core",
-    children: [
-      {
-        title: TableTitle,
-        href: TableHref,
-      },
-      {
-        title: PaginationTitle,
-        href: PaginationHref,
-      },
-      {
-        title: ToastTitle,
-        href: ToastHref,
-      },
-      {
-        title: CheckboxTitle,
-        href: CheckboxHref,
-      },
-      {
-        title: DialogTitle,
-        href: DialogHref,
-      },
-      {
-        title: SwitchTitle,
-        href: SwitchHref,
-      },
-      {
-        title: TooltipTitle,
-        href: TooltipHref,
-      },
-      {
-        title: "Template",
-        href: "template",
-      },
-    ],
-  },
-];
