+++ packages/common/index.ts
@@ -1,4 +1,4 @@
 // Barrel file
 
 // Common Multi Components
-export * from "./Json";
+export * from "./Json/Component_Matcher";
