--- packages/components/core/Collapsible/usage.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { Collapsible } from "./src";
-
-type Props = {};
-
-export const Usage = (props: Props) => {
-  return <div className=""></div>;
-};
