--- packages/components/core/DialogBox/code_text.tsx
@@ -1,20 +1,0 @@
-import React from "react";
-
-type Props = {};
-
-export const code = `
-"use client";
-import React from "react";
-import {
-  motion,
-  useAnimationFrame,
-  useMotionTemplate,
-  useMotionValue,
-  useTransform,
-} from "framer-motion";
-import { useRef } from "react";
-import { cn } from "@/lib/utils";
-
-
-
-`;
