+++ packages/components/core/DialogBox/variants/index.ts
