+++ packages/components/core/Pagination/variants/index.ts
