+++ packages/components/core/Switch/index.tsx
@@ -1,10 +1,14 @@
-// Exports for Rendering Individual Component
-export * from "./code_text";
-export * from "./description";
-export * from "./src";
-export * from "./usage";
+import React from "react";
+import { Switch } from "./src";
 
-// Exports for Reference
-export const Title = `Switch`;
+type Props = {};
 
-export const href = `switch`;
+const Switch_Usage = (props: Props) => {
+  return (
+    <div className="w-24 h-10">
+      <Switch />
+    </div>
+  );
+};
+
+export default Switch_Usage;
