--- packages/components/core/Switch/usage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import { Switch } from "./src";
-
-type Props = {};
-
-export const Usage = (props: Props) => {
-  return (
-    <div className="w-24 h-10">
-      <Switch />
-    </div>
-  );
-};
