+++ packages/components/core/Switch/variants/index.ts
