+++ packages/components/core/Table/variants/index.ts
