--- packages/components/core/__template/variants/index.ts
@@ -1,5 +1,0 @@
-export const Description = `This is a description.`;
-
-export const Props = {
-  prop: "prop",
-};
